Added player bodies migration
Added migration to hold info for player question bodies. Also note that
the name of the migration itself can't bemore than 32 characters.

I'm thining that the structure of the questions will be something like
this:

Title	The Question Title
p1_body	The question for user 1, (Write function to...)
p2_body The same thing as p1_body but for user 2