Added answers table to database.
Be sure to run migration script to update database to newest version.
Added answers table that will contain a reference to it's parent
question as well as a series of answers that goes with that question.
Which we could store as a serialized array or something simple like
that. That way when we do comparisons between what the javascript evals
to, we can just send it off and go.