Added Initial Questions API
JSON response for querying api/questions.php
Simply outputs the database to JSON with an extra field of how many rows
there were returned.

Created config file for credentials and updated migrate to use it.

Next steps:
 - make javascript consume the resource
 - .htaccess files for nice urls?
 - Think about how to create a messaging system to relay traffic to
   users using the system. Cookies may assist the client here.

Side Thoughts:
 For messaging I'm envisioning a simple table that has, essentially, a
version of the pubnub protocol I came up with previously. So there would
be a TO FROM and a message of some kind that would assist with syncing
the states of the users. The big thing is that this information is
rather transient and doesn't actually need to be saved. SO! I'm thinking
we use a Memory Engine table that way the messages are flushed out, and
perhaps have a simple cron job that kills off messages that are older
than an hour. (include timestamp in database schema for messages)